[UPnP] functioning messages [...]
I was going to add a feature that said what your current external IP was in stdout for the Opening... message (previously was meant to be Redirecting %ip:%port to %externalip), but attempting to print the externalip via UPNP_GetExternalIPAddress kept SIGSEGV'ing on me so I just abandoned it.

Oh well. next thing from here is to allow servers listening to 0.0.0.0 to function with UPnP, and not have to set the listen to your router-side IP, blocking localhost/loopback connections from functioning.